fix(usage): make the opencode scraper compile
`top_level_number` answers an f64 while the block returns `resetInSec` as seconds,
so the value is converted. The previous commit reached staging without this,
because the fix was applied after the check and the commit carried the unchecked
file anyway.

+++ apps/desktop/src-tauri/src/usage.rs
@@ -789,8 +789,6 @@ async fn kimi_usage() -> Option<AgentUsage> {
     })
 }
 
-/// @notice Reads the usage limits of one agent CLI.
-/// @dev Asked per agent, because only a few providers publish an account limit —
 // ---------------------------------------------------------------------------
 // Web sessions (MiniMax, opencode)
 // ---------------------------------------------------------------------------
@@ -993,7 +991,7 @@ fn opencode_block(text: &str, key: &str) -> Option<(f64, i64)> {
             top_level_number(block, "usagePercent"),
             top_level_number(block, "resetInSec"),
         ) {
-            return Some((percent, reset));
+            return Some((percent, reset as i64));
         }
     }
     None
